# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vapid-public-key:
    get:
      operationId: getVapidPublicKey
      tags: [notifications]
      summary: Get VAPID public key for web push subscription
      responses:
        "200":
          description: VAPID public key
          content:
            application/json:
              schema:
                type: object
                properties:
                  publicKey:
                    type: string
                required:
                  - publicKey
        "503":
          description: Web push not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notification history
      parameters:
        - name: roomNumber
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send a push notification to guests (staff only)
      security:
        - staffToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationRequest"
      responses:
        "201":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized — invalid staff token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notifications/staff-subscribe:
    post:
      operationId: registerStaffPushSubscription
      tags: [notifications]
      summary: Register staff browser push subscription (staff only)
      security:
        - staffToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffPushSubscribeRequest"
      responses:
        "201":
          description: Subscription registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unregisterStaffPushSubscription
      tags: [notifications]
      summary: Unregister staff browser push subscription (staff only)
      security:
        - staffToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffPushUnsubscribeRequest"
      responses:
        "200":
          description: Subscription removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    staffToken:
      type: http
      scheme: bearer
  schemas:
...
    SendNotificationRequest:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [room_ready, activity, checkout_reminder, general]
        targetRoom:
          type: string
          description: "Room number to target, or 'all' for all guests"
      required:
        - title
        - body
        - type
        - targetRoom
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        type:
          type: string
        targetRoom:
          type: string
        sentAt:
          type: string
        recipientCount:
          type: integer
        sentByName:
          type: string
          nullable: true
      required:
        - id
        - title
        - body
        - type
        - targetRoom
        - sentAt
        - recipientCount
...
    StaffPushSubscribeRequest:
      type: object
      properties:
        subscription:
          type: object
          description: Web Push subscription object (from PushManager.subscribe)
          additionalProperties: true
      required:
        - subscription
    StaffPushUnsubscribeRequest:
      type: object
      properties:
        endpoint:
          type: string
          description: The endpoint URL of the subscription to remove
      required:
        - endpoint
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error